+++ lib/bane.rb
@@ -12,9 +12,9 @@ require 'bane/behaviors/responders/newline_response'
 require 'bane/behaviors/responders/random_response'
 require 'bane/behaviors/responders/slow_response'
 require 'bane/behaviors/responders/exported'
-require 'bane/services/responder_server'
-require 'bane/services/timeout_in_listen_queue'
-require 'bane/services/exported'
+require 'bane/behaviors/services/responder_server'
+require 'bane/behaviors/services/timeout_in_listen_queue'
+require 'bane/behaviors/services/exported'
 require 'bane/launcher'
 require 'bane/arguments_parser'
 require 'bane/command_line_configuration'
