+++ lib/bane/behaviors/services/exported.rb
@@ -1,0 +1,19 @@
+module Bane
+  module Behaviors
+    module Services
+
+      EXPORTED = [TimeoutInListenQueue]
+
+      # Listen only on localhost
+      LOCALHOST = '127.0.0.1'
+
+      # Deprecated - use LOCALHOST - Listen only on localhost
+      DEFAULT_HOST = LOCALHOST
+
+      # Listen on all interfaces
+      ALL_INTERFACES = '0.0.0.0'
+
+
+    end
+  end
+end
